bugfixes, added availale pids command for mode 01

+++ src/main/java/pt/lighthouselabs/obd/commands/control/DistanceTraveledSinceCodesClearedObdCommand.java
@@ -49,7 +49,7 @@ public class DistanceTraveledSinceCodesClearedObdCommand extends ObdCommand
  
   public String getFormattedResult() {
     return useImperialUnits ? String.format("%.2f%s", getImperialUnit(), getResultUnit())
-            : String.format("%d%s", (float)km, getResultUnit());
+            : String.format("%d%s", km, getResultUnit());
   }
